// src/app/models/vendor-model/vendor-model.ts
export class VendorModel {
    id: string;
    code: string;
    name: string;
    companyName: string;
    description: string;
    logo: string;
    phone: string;
    fax: string;
    email: string;
    website: string;
    type: number;
    typeName: string;
    status: number;
    statusName: string;
}